backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pypdf2>=3.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
pydantic>=2.0.0
aiofiles>=23.2.0
requests>=2.31.0
